Refactor code to be less complex.
I've talked with some people on #lisp and learned that building Common
Lisp is actually very hard.

I've decided that Bleso will be its own "dialect" of Lisp, which will be
fully pure and functional, with some sugar for things like `let`
etcetera.

This is totally a stylistic choice and not a choice based on not wanting
to be bothered constructing mutability in Rust.